Move and harden reverse sound project page
